iter5 рефакторинг
