+++ internal/app/handlers/save_new_url.go
@@ -1,0 +1,30 @@
+package handlers
+
+import (
+	"github.com/KraDM09/shortener/internal/app/config"
+	"github.com/KraDM09/shortener/internal/app/storage"
+	"github.com/KraDM09/shortener/internal/app/util"
+	"io"
+	"net/http"
+)
+
+func SaveNewURLHandler(rw http.ResponseWriter, r *http.Request, store storage.Storage) {
+	body, err := io.ReadAll(r.Body)
+
+	if err != nil {
+		http.Error(rw, "Ошибка чтения тела запроса", http.StatusBadRequest)
+		return
+	}
+
+	URL := string(body)
+
+	hash := util.CreateHash()
+	store.Save(hash, URL)
+
+	rw.Header().Set("Content-Type", "text/plain")
+	rw.WriteHeader(http.StatusCreated)
+	_, err = rw.Write([]byte(config.FlagBaseShortURL + "/" + hash))
+	if err != nil {
+		panic(err)
+	}
+}
